requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.3
pyyaml>=6.0.1
pygithub>=2.3.0
arxiv>=1.4.8
feedparser>=6.0.11
python-dotenv>=1.0.1